add api controller for make test routes
